+++ src/compas_rhino/geometry/brep/__init__.py
@@ -76,6 +76,11 @@ def from_plane(*args, **kwargs):
 
 
 @plugin(category="factories", requires=["Rhino"])
+def from_polygons(*args, **kwargs):
+    return RhinoBrep.from_polygons(*args, **kwargs)
+
+
+@plugin(category="factories", requires=["Rhino"])
 def from_sphere(*args, **kwargs):
     return RhinoBrep.from_sphere(*args, **kwargs)